Consolidate ACKNACK related code
